introduce general type

+++ deepqmc/solver/solver_base.py
@@ -40,7 +40,7 @@ class SolverBase(object):
             self.wf.pdf, ntherm=ntherm, ndecor=ndecor,
             with_tqdm=with_tqdm, pos=pos)
         pos.requires_grad = True
-        return pos.float()
+        return pos
 
     def get_observable(self, obs_dict, pos, **kwargs):
         '''compute all the required observable.
